+++ SofaKernel/modules/SofaBaseTopology/QuadSetGeometryAlgorithms.inl
@@ -106,7 +106,7 @@ template<class DataTypes>
 void QuadSetGeometryAlgorithms<DataTypes>::computeQuadArea( BasicArrayInterface<Real> &ai) const
 {
     //const sofa::helper::vector<Quad> &ta=this->m_topology->getQuads();
-    int nb_quads = this->m_topology->getNbQuads();
+    size_t nb_quads = this->m_topology->getNbQuads();
     const typename DataTypes::VecCoord& p =(this->object->read(core::ConstVecCoordId::position())->getValue());
 
     for(int i=0; i<nb_quads; ++i)
